- put message sender and recorders into actors to isolate meeting actor. This hopefully    will isolate the meeting actor from exceptions when sending to pusub and recording to redis
