+++ akka-bbb-apps/src/main/scala/org/bigbluebutton/Boot.scala
@@ -18,10 +18,6 @@ import org.bigbluebutton.core.pubsub.senders._
 import org.bigbluebutton.core.service.recorder.RedisDispatcher
 import org.bigbluebutton.core.service.recorder.RecorderApplication
 import org.bigbluebutton.core.recorders.VoiceEventRecorder
-import org.bigbluebutton.core.recorders.ChatEventRedisRecorder
-import org.bigbluebutton.core.recorders.PresentationEventRedisRecorder
-import org.bigbluebutton.core.recorders.UsersEventRedisRecorder
-import org.bigbluebutton.core.recorders.WhiteboardEventRedisRecorder
 
 object Boot extends App with SystemConfiguration {
 
@@ -35,33 +31,7 @@ object Boot extends App with SystemConfiguration {
   recorderApp.start()
 
   val voiceEventRecorder = new VoiceEventRecorder(recorderApp)
-
-  val chatEventRecorder = new ChatEventRedisRecorder(recorderApp)
-  val presentationEventRecorder = new PresentationEventRedisRecorder(recorderApp)
-  val usersEventRecorder = new UsersEventRedisRecorder(recorderApp)
-  val whiteboardEventRecorder = new WhiteboardEventRedisRecorder(recorderApp)
-
-  val chatSender = new ChatEventRedisPublisher(msgSender)
-  val meetingSender = new MeetingEventRedisPublisher(msgSender)
-  val presSender = new PresentationEventRedisPublisher(msgSender)
-  val userSender = new UsersEventRedisPublisher(msgSender)
-  val whiteboardSender = new WhiteboardEventRedisPublisher(msgSender)
-  val pollingSender = new PollingEventRedisPublisher(msgSender)
-
-  val outMessageListeners = new java.util.ArrayList[OutMessageListener2]()
-  outMessageListeners.add(chatSender)
-  outMessageListeners.add(meetingSender)
-  outMessageListeners.add(presSender)
-  outMessageListeners.add(userSender)
-  outMessageListeners.add(whiteboardSender)
-  outMessageListeners.add(chatEventRecorder)
-  outMessageListeners.add(presentationEventRecorder)
-  outMessageListeners.add(usersEventRecorder)
-  outMessageListeners.add(whiteboardEventRecorder)
-  outMessageListeners.add(pollingSender)
-
-  val outGW = new MessageOutGateway(outMessageListeners)
-  val bbbInGW = new BigBlueButtonInGW(system, outGW, voiceEventRecorder)
+  val bbbInGW = new BigBlueButtonInGW(system, recorderApp, msgSender, voiceEventRecorder)
   val redisMsgReceiver = new RedisMessageReceiver(bbbInGW)
 
   val redisSubscriberActor = system.actorOf(AppsRedisSubscriberActor.props(redisMsgReceiver), "redis-subscriber")
