+++ akka-bbb-apps/src/main/scala/org/bigbluebutton/core/RunningMeeting.scala
@@ -5,13 +5,11 @@ import akka.actor.ActorContext
 import org.bigbluebutton.core.api.MessageOutGateway
 
 object RunningMeeting {
-  def apply(mProps: MeetingProperties,
-    outGW: MessageOutGateway)(implicit context: ActorContext) =
+  def apply(mProps: MeetingProperties, outGW: OutMessageGateway)(implicit context: ActorContext) =
     new RunningMeeting(mProps, outGW)(context)
 }
 
-class RunningMeeting(val mProps: MeetingProperties,
-    val outGW: MessageOutGateway)(implicit val context: ActorContext) {
+class RunningMeeting(val mProps: MeetingProperties, val outGW: OutMessageGateway)(implicit val context: ActorContext) {
 
   val actorRef = context.actorOf(MeetingActor.props(mProps, outGW), mProps.meetingID)
 }
