+++ akka-bbb-apps/src/main/scala/org/bigbluebutton/core/apps/ChatApp.scala
@@ -3,11 +3,12 @@ package org.bigbluebutton.core.apps
 import org.bigbluebutton.core.api._
 import scala.collection.mutable.ArrayBuffer
 import org.bigbluebutton.core.MeetingActor
+import org.bigbluebutton.core.OutMessageGateway
 
 trait ChatApp {
   this: MeetingActor =>
 
-  val outGW: MessageOutGateway
+  val outGW: OutMessageGateway
 
   def handleGetChatHistoryRequest(msg: GetChatHistoryRequest) {
     val history = chatModel.getChatHistory()
