--- akka-bbb-apps/src/main/scala/org/bigbluebutton/core/pubsub/senders/MeetingEventRedisPublisher.scala
@@ -1,110 +1,0 @@
-package org.bigbluebutton.core.pubsub.senders
-
-import org.bigbluebutton.core.MessageSender
-import org.bigbluebutton.core.api._
-import scala.collection.immutable.HashMap
-import com.google.gson.Gson
-import scala.collection.JavaConverters._
-import org.bigbluebutton.common.messages.MessagingConstants
-import org.bigbluebutton.common.messages.StartRecordingVoiceConfRequestMessage
-import org.bigbluebutton.common.messages.StopRecordingVoiceConfRequestMessage
-
-class MeetingEventRedisPublisher(service: MessageSender) extends OutMessageListener2 {
-
-  private val KEEP_ALIVE_REPLY = "KEEP_ALIVE_REPLY"
-
-  def handleMessage(msg: IOutMessage) {
-    msg match {
-      case msg: MeetingCreated => handleMeetingCreated(msg)
-      case msg: VoiceRecordingStarted => handleVoiceRecordingStarted(msg)
-      case msg: VoiceRecordingStopped => handleVoiceRecordingStopped(msg)
-      case msg: RecordingStatusChanged => handleRecordingStatusChanged(msg)
-      case msg: GetRecordingStatusReply => handleGetRecordingStatusReply(msg)
-      case msg: MeetingEnded => handleMeetingEnded(msg)
-      case msg: MeetingHasEnded => handleMeetingHasEnded(msg)
-      case msg: MeetingDestroyed => handleMeetingDestroyed(msg)
-      case msg: KeepAliveMessageReply => handleKeepAliveMessageReply(msg)
-      case msg: StartRecording => handleStartRecording(msg)
-      case msg: StopRecording => handleStopRecording(msg)
-      case msg: GetAllMeetingsReply => handleGetAllMeetingsReply(msg)
-      case msg: StartRecordingVoiceConf => handleStartRecordingVoiceConf(msg)
-      case msg: StopRecordingVoiceConf => handleStopRecordingVoiceConf(msg)
-      case _ => //println("Unhandled message in MeetingEventRedisPublisher")
-    }
-  }
-
-  private def handleStartRecordingVoiceConf(msg: StartRecordingVoiceConf) {
-    val m = new StartRecordingVoiceConfRequestMessage(msg.meetingID, msg.voiceConfId)
-    System.out.println("*******************\n" + m.toJson())
-    service.send(MessagingConstants.TO_VOICE_CONF_SYSTEM_CHAN, m.toJson())
-  }
-
-  private def handleStopRecordingVoiceConf(msg: StopRecordingVoiceConf) {
-    val m = new StopRecordingVoiceConfRequestMessage(msg.meetingID, msg.voiceConfId, msg.recordedStream)
-    System.out.println("*******************\n" + m.toJson())
-    service.send(MessagingConstants.TO_VOICE_CONF_SYSTEM_CHAN, m.toJson())
-  }
-
-  private def handleMeetingDestroyed(msg: MeetingDestroyed) {
-    val json = MeetingMessageToJsonConverter.meetingDestroyedToJson(msg)
-    //    System.out.println("****\n" + json)
-    service.send(MessagingConstants.FROM_MEETING_CHANNEL, json)
-  }
-
-  private def handleKeepAliveMessageReply(msg: KeepAliveMessageReply) {
-    val json = MeetingMessageToJsonConverter.keepAliveMessageReplyToJson(msg)
-    service.send(MessagingConstants.FROM_SYSTEM_CHANNEL, json)
-  }
-
-  private def handleMeetingCreated(msg: MeetingCreated) {
-    val json = MeetingMessageToJsonConverter.meetingCreatedToJson(msg)
-    //    System.out.println("****\n" + json)
-    service.send(MessagingConstants.FROM_MEETING_CHANNEL, json)
-  }
-
-  private def handleMeetingEnded(msg: MeetingEnded) {
-    val json = MeetingMessageToJsonConverter.meetingEndedToJson(msg)
-    //    System.out.println("****\n" + json)
-    service.send(MessagingConstants.FROM_MEETING_CHANNEL, json)
-  }
-
-  private def handleStartRecording(msg: StartRecording) {
-    val json = MeetingMessageToJsonConverter.startRecordingToJson(msg)
-    service.send(MessagingConstants.FROM_MEETING_CHANNEL, json)
-  }
-
-  private def handleStopRecording(msg: StopRecording) {
-    val json = MeetingMessageToJsonConverter.stopRecordingToJson(msg)
-    service.send(MessagingConstants.FROM_MEETING_CHANNEL, json)
-  }
-
-  private def handleVoiceRecordingStarted(msg: VoiceRecordingStarted) {
-    val json = MeetingMessageToJsonConverter.voiceRecordingStartedToJson(msg)
-    service.send(MessagingConstants.FROM_MEETING_CHANNEL, json)
-  }
-
-  private def handleVoiceRecordingStopped(msg: VoiceRecordingStopped) {
-    val json = MeetingMessageToJsonConverter.voiceRecordingStoppedToJson(msg)
-    service.send(MessagingConstants.FROM_MEETING_CHANNEL, json)
-  }
-
-  private def handleRecordingStatusChanged(msg: RecordingStatusChanged) {
-    val json = MeetingMessageToJsonConverter.recordingStatusChangedToJson(msg)
-    service.send(MessagingConstants.FROM_MEETING_CHANNEL, json)
-  }
-
-  private def handleGetRecordingStatusReply(msg: GetRecordingStatusReply) {
-    val json = MeetingMessageToJsonConverter.getRecordingStatusReplyToJson(msg)
-    service.send(MessagingConstants.FROM_MEETING_CHANNEL, json)
-  }
-
-  private def handleMeetingHasEnded(msg: MeetingHasEnded) {
-    val json = MeetingMessageToJsonConverter.meetingHasEndedToJson(msg)
-    service.send(MessagingConstants.FROM_MEETING_CHANNEL, json)
-  }
-
-  private def handleGetAllMeetingsReply(msg: GetAllMeetingsReply) {
-    val json = MeetingMessageToJsonConverter.getAllMeetingsReplyToJson(msg)
-    service.send(MessagingConstants.FROM_MEETING_CHANNEL, json)
-  }
-}
